+++ sharding-jdbc/sharding-jdbc-core/src/main/java/io/shardingsphere/shardingjdbc/jdbc/adapter/AbstractDataSourceAdapter.java
@@ -23,7 +23,7 @@ import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.util.ReflectiveUtil;
 import io.shardingsphere.shardingjdbc.jdbc.unsupported.AbstractUnsupportedOperationDataSource;
 import io.shardingsphere.transaction.spi.xa.DataSourceMapConverter;
-import io.shardingsphere.transaction.api.SPIDataSourceMapConverter;
+import io.shardingsphere.transaction.core.api.SPIDataSourceMapConverter;
 import lombok.Getter;
 import lombok.Setter;
